requirements.txt
fastapi
pydantic
python-dotenv
uvicorn
Pillow
ultralytics
onnxruntime
Rembg